+++ app/src/docs/Rating.doc.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
+import * as uui from '@epam/uui';
 import * as loveship from '@epam/loveship';
 import * as promo from '@epam/promo';
 import { DocBuilder, TDocConfig, TDocContext, TSkin } from '@epam/uui-docs';
 import { BaseDocsBlock, DocExample, EditableDocContent } from '../common';
-import * as uui from '@epam/uui';
 
 export class RatingDoc extends BaseDocsBlock {
     title = 'Rating';
@@ -12,6 +12,7 @@ export class RatingDoc extends BaseDocsBlock {
         name: 'Rating',
         contexts: [TDocContext.Default, TDocContext.Form],
         bySkin: {
+            [TSkin.UUI]: { type: '@epam/uui:RatingProps', component: uui.Rating },
             [TSkin.Loveship]: { type: '@epam/loveship:RatingProps', component: loveship.Rating },
             [TSkin.Promo]: { type: '@epam/promo:RatingProps', component: promo.Rating },
         },
@@ -23,6 +24,7 @@ export class RatingDoc extends BaseDocsBlock {
                     0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5,
                 ],
             });
+            doc.merge('hint', { examples: [{ value: (rating) => `${rating} star(s)`, name: '(rating) => string' }] });
         },
     };